Restructure/Add some components
