+++ packages/react-sdk-components/src/components/infra/VerticalTabs/LeftAlignVerticalTabs/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './LeftAlignVerticalTabs';
