+++ packages/react-sdk-components/src/components/infra/VerticalTabs/VerticalTabs/index.tsx
