+++ packages/react-sdk-components/src/components/templates/CaseViewActionsMenu/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './CaseViewActionsMenu';
