+++ packages/react-sdk-components/src/components/widgets/FileUtility/FileUtility/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './FileUtility';
