+++ scripts/override-constants.js
@@ -16,6 +16,7 @@ const sdkTopLevelContent = [ 'components_map' ];
 //    'simpleTableHelpers' is in templates/SimpleTable' directory (a non-component entry in this table)
 const sdkComponentLocationMap = {
   'ActionButtons': 'infra',
+  'ActionButtonsForFileUtil': 'widgets/FileUtility',
   'AppAnnouncement': 'widgets',
   'AppShell': 'templates',
   'Assignment': 'infra',
@@ -27,6 +28,7 @@ const sdkComponentLocationMap = {
   'CaseSummary': 'templates',
   'CaseSummaryFields': 'designSystemExtensions',
   'CaseView': 'templates',
+  'CaseViewActionsMenu': 'templates',
   'Checkbox': 'forms',
   'Currency': 'forms',
   'DashboardFilter': 'infra',
@@ -44,12 +46,13 @@ const sdkComponentLocationMap = {
   'Email': 'forms',
   'ErrorBoundary': 'infra',
   'FieldGroupTemplate': 'templates',
-  'FileUtility': 'widgets',
+  'FileUtility': 'widgets/FileUtility',
   'FlowContainer': 'infra',
   'Followers': 'widgets',
   'InlineDashboard': 'templates',
   'InlineDashboardPage': 'templates',
   'Integer': 'forms',
+  'LeftAlignVerticalTabs': 'infra/VerticalTabs',
   'ListPage': 'templates',
   'ListView': 'templates',
   'ModalViewContainer': 'infra',
@@ -92,7 +95,7 @@ const sdkComponentLocationMap = {
   'TwoColumnTab': 'templates',
   'URL': 'forms',
   'UserReference': 'forms',
-  'VerticalTabs': 'infra',
+  'VerticalTabs': 'infra/VerticalTabs',
   'View': 'infra',
   'ViewContainer': 'infra',
   'WideNarrow': 'templates',
